No file found error

+++ src/index.ts
@@ -1,16 +1,13 @@
 import * as fs from 'fs';
-import { Generator } from "./generator"
+if (!fs.existsSync("./skript/src/index.ts")) {
+	console.error('The file "./skript/src/index.ts" could not be found!');
+	process.exit(1);
+}
+import { Generator } from "./generator";
 
-
-
-
-
-let gen = new Generator(fs.readFileSync("./skript/src/index.ts").toString())
+let gen = new Generator(fs.readFileSync("./skript/src/index.ts").toString());
 gen.parse((skript: string) => {
 	fs.writeFile("./skript/out/index.sk", skript, (err) => {
 		if (err) console.log(err);
-	  })
-})
-
-
-
+	});
+});
